speak_words/requirements.txt
fastapi
uvicorn
boto3
google-auth
google-auth-oauthlib
google-cloud-texttospeech
